Add more SDK data structure

+++ sdk/gocqhttp/data/General/SendGroupMsg.go
@@ -2,7 +2,7 @@ package General
 
 // SendGroupMsg Endpoint send_group_msg
 type SendGroupMsg struct {
-	GroupId    int64  `json:"group_id"`
+	GroupId    int    `json:"group_id"`
 	Message    string `json:"message"`
 	AutoEscape bool   `json:"auto_escape"`
 }
